+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-picker/example/DatePickerDefault.jsx
@@ -8,14 +8,16 @@ const cx = classNames.bind(styles);
 
 const DatePickerExampleDefault = () => {
   const [date, setDate] = useState('');
-
+  const [invalid, setInValid] = useState(false);
   const handleDateChange = (event, dateValue) => {
     setDate(dateValue);
+    setInValid(false);
   };
 
   const handleDateChangeRaw = (event, dateValue, metadata) => {
-    if (!metadata.isValidValue) {
+    if (!metadata.isValidValue && metadata.inputValue.length === 10) {
       setDate(null);
+      setInValid(true);
     }
   };
 
@@ -31,6 +33,7 @@ const DatePickerExampleDefault = () => {
           id="default"
           onChange={handleDateChange}
           onChangeRaw={handleDateChangeRaw}
+          isInvalid={invalid}
         />
       </Field>
     </div>